claudey/ai_model/requirements.txt
fastapi
uvicorn
transformers
accelerate
bitsandbytes
python-dotenv
torch --index-url https://download.pytorch.org/whl/cpu